Rename `FormLabel` component to `Label`
Need to come back to the documentation later
