+++ examples/create-react-app-typescript/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@cmsgov/design-system": "^6.0.1",
+    "@cmsgov/design-system": "^8.0.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.0",
     "@types/react": "^18.0.22",
