+++ examples/create-react-app-typescript/src/components/Examples/LabelExample.tsx
@@ -1,20 +1,20 @@
-import { FormLabel } from '@cmsgov/design-system';
+import { Label } from '@cmsgov/design-system';
 import React from 'react';
 
-function FormLabelExample(): React.ReactElement {
+function LabelExample(): React.ReactElement {
   return (
     <div>
-      <h2>FormLabel Example</h2>
-      <FormLabel
+      <h2>Label Example</h2>
+      <Label
         errorMessage="Please enter a valid phone number."
         hint="We'll use this number as a backup if we need to contact you about your application."
         fieldId="number-field-id"
         requirementLabel="Optional"
       >
         Phone number
-      </FormLabel>
+      </Label>
     </div>
   );
 }
 
-export default FormLabelExample;
+export default LabelExample;
