+++ packages/design-system/src/components/Label/index.ts
@@ -1,0 +1,1 @@
+export * from './Label';
