+++ packages/design-system/src/styles/components/_Label.scss
