+++ packages/design-system/src/styles/index.scss
@@ -48,7 +48,7 @@
 @use 'components/Drawer';
 @use 'components/Dropdown';
 @use 'components/FilterChip';
-@use 'components/FormLabel';
+@use 'components/Label';
 @use 'components/List';
 @use 'components/MonthPicker';
 @use 'components/Pagination';
