+++ packages/docs/content/utilities/focus.mdx
@@ -2,7 +2,7 @@
 title: Focus
 ---
 
-import { Button, FormLabel } from '@cmsgov/design-system';
+import { Button, Label } from '@cmsgov/design-system';
 
 Use the following utilities to apply focus styles to custom UI components.
 
@@ -17,9 +17,9 @@ Use the `.ds-u-focus` utility to conditionally apply focus styles to your compon
 Click or tab into the following example to see this utility in action.
 
 <EmbeddedExample>
-  <FormLabel fieldId="focus" className="ds-u-margin-top--0">
+  <Label fieldId="focus" className="ds-u-margin-top--0">
     Select a color
-  </FormLabel>
+  </Label>
   <input type="color" id="focus" name="color-focus" class="ds-u-focus ds-u-radius ds-u-margin-top--1" />
 </EmbeddedExample>
 
@@ -30,9 +30,9 @@ Use the `.ds-u-focus-visible` utility to conditionally apply focus styles to you
 Tab into the following example to see this utility in action.
 
 <EmbeddedExample>
-  <FormLabel fieldId="focus-visible" className="ds-u-margin-top--0">
+  <Label fieldId="focus-visible" className="ds-u-margin-top--0">
     Select another color
-  </FormLabel>
+  </Label>
   <input type="color" id="focus-visible" name="color-focus-visible" class="ds-u-focus-visible ds-u-radius ds-u-margin-top--1" />
 </EmbeddedExample>
